\name{shiny_LINES_posterior}
\alias{shiny_LINES_posterior}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
LINES prior-posterior distribution with posterior probabilities
}
\description{
This is a shiny app for the LINES trial. This trial is a dual endpoint design with both response and toxicity used to make informed decisions at interim anaylsis. This app provides an interactive way of updating the posterior distribution, as well as change the prior distributions.
}
\usage{
shiny_LINES_posterior()
}